overwrite old files with the Vue versions
